+++ Demos/DynamicLink/LinkDemo/Makefile
@@ -2,8 +2,8 @@ OBJS = \
     main.o
 
 PROGRAM = LinkDemo
-
+VERBOSE = 1
 SUBPROJECT_CXXFLAGS = -fPIC
-LDFLAGS = -pie -Wl,--dynamic-linker=/lib/ld-elf.so
+LDFLAGS = -pie -Wl,--dynamic-linker=/lib/ld-elf.so -lc -lm
 
 include ../../../Makefile.common
