+++ Libraries/LibM/Makefile
@@ -1,6 +1,13 @@
 OBJS = math.o
 
-LIBRARY = libm.a
+SHARED ?= 0
+ifeq ($(SHARED),1)
+    PROGRAM = libm.so
+    LDFLAGS = -shared
+    SUBPROJECT_CXXFLAGS = -fPIC
+else
+    LIBRARY = libm.a
+endif
 
 install:
 	mkdir -p $(SERENITY_BASE_DIR)/Root/usr/include/sys/
